backend/requirements.txt
# Flask Web Framework
Flask>=2.3.0
Flask-CORS>=4.0.0
dash>=2.14.0
dash-bootstrap-components>=1.5.0

# Data Processing
pandas>=2.0.0
numpy>=1.25.0
openpyxl>=3.0.0
dask>=2023.9.0
dask[dataframe]>=2023.9.0
vaex>=4.16.0

# Advanced NLP
spacy>=3.7.0
nltk>=3.8.0
textblob>=0.17.0
wordcloud>=1.9.0

# Visualization
matplotlib>=3.7.0
plotly>=5.15.0
seaborn>=0.12.0
bokeh>=3.2.0
altair>=5.1.0

# File Processing
PyPDF2>=3.0.0
Pillow>=10.0.0

# AI/ML
google-generativeai>=0.3.0
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0
statsmodels>=0.14.0

# Time Series Analysis
fbprophet>=0.7.1
statsforecast>=1.5.0

# Utilities
requests>=2.31.0
werkzeug>=2.3.0
kaleido
python-dotenv>=1.0.0
loguru>=0.7.0
tqdm>=4.66.0
joblib>=1.3.0
cachetools>=5.3.0

